# [wasm] Optimize jiterpreter type cast opcodes & add zero page optimizations