… src/mono/wasm/runtime/jiterpreter.ts