This PR inlines a sizable portion of MINT_CASTCLASS/MINT_ISINST logic into jiterpreter traces to create fast paths for exact matches/null pointers that avoids running all the type check machinery we normally would run. Based on my instrumentation these fast paths will be used ~30% of the time, and measurements show a 1-3% improvement for the json section of browser-bench (this is pretty hard to measure, I may try to construct a decent synthetic benchmark for this).

This PR also applies zero page optimizations if available, by fusing the null check into the `obj->vtable` load. The helper code is all tweaked so that in the event a check fails due to a null ptr (null inputs are fairly uncommon in the instrumentation) it is converted into a success since MINT_CASTCLASS and MINT_ISINST both succeed for null instead of throwing.

It may be worth inlining the interface check entirely, but I haven't done the work to test that yet. The existence of special interfaces for arrays makes that harder to do.

This PR also fully inlines the implementation of MINT_UNBOX.