@vargaz 2023-05-31
isinst doesn't succeed on nulls.