@kg 2023-05-31
> isinst doesn't succeed on nulls.

Thanks, I'll revise it