@kg 2023-06-01
I was puzzled that none of the scenarios I expected to get faster were getting faster with these changes. It turned out that the scenarios I was testing were actually relying on MINT_UNBOX. So I wrote some toy BDN microbenchmarks to use and also optimized the unbox operation. Some comparison timings for main vs this PR:

|scenario|main|this PR|relative %|
|-:|-:|-:|-:|
|`(int)obj`|1.285ms|1.159ms|**90%**|
|`if (obj is int value)`|1.62ms|1.218ms|**75%**|